+++ public_data/models/urbanisme/LogementVacant.py
@@ -11,9 +11,9 @@ class LogementVacant(models.Model):
     logements_vacants_parc_prive = models.IntegerField()
     logements_parc_social = models.IntegerField()
     logements_vacants_parc_social = models.IntegerField()
-    logements_vacants_parc_general_percent = models.DecimalField()
-    logements_vacants_parc_prive_percent = models.DecimalField()
-    logements_vacants_parc_social_percent = models.DecimalField()
+    logements_vacants_parc_general_percent = models.FloatField()
+    logements_vacants_parc_prive_percent = models.FloatField()
+    logements_vacants_parc_social_percent = models.FloatField()
 
     class Meta:
         managed = False
